Lift MAX_32 limit

+++ src/rebloom.c
@@ -124,7 +124,7 @@ static int BFReserve_RedisCommand(RedisModuleCtx *ctx, RedisModuleString **argv,
 
     long long capacity;
     if (RedisModule_StringToLongLong(argv[3], &capacity) != REDISMODULE_OK ||
-        capacity >= UINT32_MAX) {
+        capacity >= UINT64_MAX) {
         return RedisModule_ReplyWithError(ctx, "ERR bad capacity");
     }
